requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
einops==0.8.0
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.1
frozenlist==1.4.1
fsspec==2024.6.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.3
idna==3.7
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
orjson==3.10.4
packaging==24.1
pillow==10.3.0
plumbum==1.8.3
protobuf==5.27.1
pybind11==2.12.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
pyzmq==26.0.3
ray==2.24.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rpds-py==0.18.1
rpyc==6.0.0
safetensors==0.4.3
sentencepiece==0.2.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
sympy==1.12.1
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.4
transformers==4.41.2
--extra-index-url https://aiinfra.pkgs.visualstudio.com/PublicPackages/_packaging/Triton-Nightly/pypi/simple/
triton-nightly
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
pandas==2.2.2
notebook==7.2.1